refactor: 도메인이 main인 것은 너무 추상적이다. main을 event 도메인으로 이동
