+++ src/main/java/com/hyundai/softeer/backend/domain/event/dto/MainLandDto.java
@@ -1,4 +1,4 @@
-package com.hyundai.softeer.backend.domain.main.dto;
+package com.hyundai.softeer.backend.domain.event.dto;
 
 import lombok.Builder;
 import lombok.Getter;
