+++ src/main/java/com/hyundai/softeer/backend/domain/event/service/MainService.java
@@ -1,9 +1,9 @@
-package com.hyundai.softeer.backend.domain.main.service;
+package com.hyundai.softeer.backend.domain.event.service;
 
 import com.hyundai.softeer.backend.domain.event.entity.Event;
 import com.hyundai.softeer.backend.domain.event.exception.EventNotFoundException;
 import com.hyundai.softeer.backend.domain.event.repository.EventRepository;
-import com.hyundai.softeer.backend.domain.main.dto.MainLandDto;
+import com.hyundai.softeer.backend.domain.event.dto.MainLandDto;
 import com.hyundai.softeer.backend.domain.subevent.entity.SubEvent;
 import com.hyundai.softeer.backend.domain.subevent.exception.SubEventNotFoundException;
 import com.hyundai.softeer.backend.domain.subevent.repository.SubEventRepository;
