+++ src/test/java/com/hyundai/softeer/backend/domain/main/service/MainServiceTest.java
@@ -3,7 +3,8 @@ package com.hyundai.softeer.backend.domain.main.service;
 import com.hyundai.softeer.backend.domain.event.entity.Event;
 import com.hyundai.softeer.backend.domain.event.exception.EventNotFoundException;
 import com.hyundai.softeer.backend.domain.event.repository.EventRepository;
-import com.hyundai.softeer.backend.domain.main.dto.MainLandDto;
+import com.hyundai.softeer.backend.domain.event.dto.MainLandDto;
+import com.hyundai.softeer.backend.domain.event.service.MainService;
 import com.hyundai.softeer.backend.domain.subevent.entity.SubEvent;
 import com.hyundai.softeer.backend.domain.subevent.exception.SubEventNotFoundException;
 import com.hyundai.softeer.backend.domain.subevent.repository.SubEventRepository;
